fix: add deletion authority
- add deletion authority to make sure that one create is the one who can delete
- create get and update event repo

+++ server/src/v1/services/event.service.ts
@@ -2,6 +2,7 @@ import eventRepository from '../repositories/event.repository';
 import { AddEventInterface } from '../interface/event';
 import { Prisma } from '@prisma/client';
 import { NODE_ENV } from '../../constants/app.constants';
+import { NotFoundError, ForbiddenError } from '@/utils/customErrors';
 
 const addEvent = async (event_data: AddEventInterface) => {
   try {
@@ -27,8 +28,20 @@ const addEvent = async (event_data: AddEventInterface) => {
   }
 };
 
-const deleteEvent = async (eventId: string) => {
-  return await eventRepository.deleteEvent(eventId);
+const deleteEvent = async (eventId: string, created_by: string) => {
+  try {
+    const event = await eventRepository.getEventDetails(eventId);
+    if (!event) {
+      throw new NotFoundError('Event not found');
+    }
+    if (event.created_by !== created_by) {
+      throw new ForbiddenError('You are not authorized to delete this event');
+    }
+    return await eventRepository.deleteEvent(eventId);
+  } catch (error) {
+    console.error('Delete event failed:', error);
+    throw error;
+  }
 };
 
 const eventServices = {
